gas: pack nonce and numberOfClaimsSent into single storage slot
Reorder JBOutboxTree struct so that `nonce` (uint64, 8 bytes) and
`numberOfClaimsSent` (narrowed from uint256 to uint192, 24 bytes)
share one 32-byte storage slot. Both fields are read together in
`_sendRoot()`, saving one cold SLOAD (~2.1K gas) per `toRemote` call.

uint192 max (6.27e57) is far beyond any realistic tree count since
each `prepare()` increments count by 1.

+++ src/JBSucker.sol
@@ -1065,7 +1065,7 @@ abstract contract JBSucker is ERC2771Context, JBPermissioned, Initializable, ERC
 
         // Update the numberOfClaimsSent to the current count of the tree.
         // This is used as in the fallback to allow users to withdraw locally if the bridge is reverting.
-        outbox.numberOfClaimsSent = count;
+        outbox.numberOfClaimsSent = uint192(count);
         uint256 index = count - 1;
 
         // Emit an event for the relayers to watch for.
